refactor caching using public functions
