+++ src/core.rs
@@ -1,15 +1,16 @@
 //! # rltbl/rltbl_db
 
 use crate::{
+    any::AnyPool,
     cache::{
         CachingStrategy, MemoryQueryCacheKey, MemoryQueryCacheValue, QUERY_CACHE_TABLE,
-        TABLE_CACHE_TABLE, clear_memory_query_cache, exists_in_meta_cache, get_memory_query_cache,
-        get_memory_table_cache, get_meta_cache,
+        ensure_cache_tables_exist, ensure_caching_triggers_exist_for_table,
+        ensure_caching_triggers_exist_for_view, exists_in_meta_cache, get_memory_query_cache,
+        get_meta_cache, update_cached_views, update_last_verified,
     },
     db_kind::DbKind,
-    db_value::{ColumnMap, DbParams, DbRow, DbRows, DbValue, IntoDbParams, IntoDbRows},
-    params,
-    parse::{get_accessed_tables, get_affected_tables, get_view_tables},
+    db_value::{ColumnMap, DbParams, DbRow, DbRows, IntoDbParams, IntoDbRows},
+    parse::get_accessed_tables,
 };
 
 use async_trait::async_trait;
@@ -18,7 +19,8 @@ use std::{
     collections::HashSet,
     fmt::Display,
     future::Future,
-    time::{Duration, SystemTime, UNIX_EPOCH},
+    marker::Sync,
+    time::{SystemTime, UNIX_EPOCH},
 };
 
 /// Defines the various error types that can be generated by rltbl_db.
@@ -76,7 +78,7 @@ impl std::fmt::Display for DbError {
 }
 
 #[async_trait]
-pub trait DbQuery {
+pub trait DbQuery: Sync {
     /// Set the caching strategy.
     fn set_caching_strategy(&mut self, strategy: &CachingStrategy);
 
@@ -99,458 +101,6 @@ pub trait DbQuery {
 
     /// Returns true if the cache-aware-query option is currently on.
     fn get_cache_aware_query(&self) -> bool;
-
-    /// Ensure that the query cache table and the table cache table exist (see
-    /// [QUERY_CACHE_TABLE] and [TABLE_CACHE_TABLE]).
-    async fn ensure_cache_tables_exist(&self) -> Result<(), DbError> {
-        if !exists_in_meta_cache(QUERY_CACHE_TABLE)? || !exists_in_meta_cache(TABLE_CACHE_TABLE)? {
-            for special_table in [QUERY_CACHE_TABLE, TABLE_CACHE_TABLE] {
-                let sql = match special_table {
-                    table if table == QUERY_CACHE_TABLE => {
-                        self.kind().create_query_cache_table_sql()
-                    }
-                    table if table == TABLE_CACHE_TABLE => {
-                        self.kind().create_table_cache_table_sql()
-                    }
-                    _ => unreachable!(),
-                };
-                match self.execute_no_cache_clean(&sql, ()).await {
-                    Ok(_) => (),
-                    Err(_) => {
-                        // Since we are not using transactions, a race condition could occur in
-                        // which two or more threads are trying to create the cache at the same
-                        // time, triggering a primary key violation in the metadata table. So if
-                        // there is an error creating the cache table we just check that it exists
-                        // and if it does we assume that all is ok.
-                        match self.table_exists(special_table).await? {
-                            false => {
-                                return Err(DbError::DatabaseError(format!(
-                                    "The cache table '{special_table}' could not be created"
-                                )));
-                            }
-                            true => (),
-                        }
-                    }
-                };
-                let mut cache = get_meta_cache()?;
-                cache.insert(special_table.to_string());
-            }
-        }
-        Ok(())
-    }
-
-    /// Ensure that caching triggers exist for the given table. Note that this function calls
-    /// [DbQuery::ensure_cache_tables_exist()] implicitly.
-    async fn ensure_caching_triggers_exist_for_table(&self, table: &str) -> Result<(), DbError> {
-        let table_triggers_name = format!("{table}_triggers");
-        if !exists_in_meta_cache(&table_triggers_name)? {
-            self.ensure_cache_tables_exist().await?;
-            let sql = self
-                .kind()
-                .create_table_caching_triggers_for_table_sql(&table)?
-                .join(";\n");
-            self.execute_batch(&sql).await?;
-
-            // Indicate that triggers exist for `table` in the meta-cache:
-            let mut cache = get_meta_cache()?;
-            cache.insert(table_triggers_name);
-        }
-        Ok(())
-    }
-
-    /// Ensure that caching triggers exist for the source tables of the given view. Note that
-    /// this function calls [DbQuery::ensure_cache_tables_exist()] implicitly.
-    async fn ensure_caching_triggers_exist_for_view(&self, view: &str) -> Result<(), DbError> {
-        let view_triggers_name = format!("{view}_triggers");
-        if !exists_in_meta_cache(&view_triggers_name)? {
-            self.ensure_cache_tables_exist().await?;
-            let view_sql = self.get_view_sql(&view).await?;
-            let source_tables = get_view_tables(&view_sql)?;
-            for source_table in source_tables.iter() {
-                // Add a trigger to clean entries from the cache for the source table itself:
-                let sql = self
-                    .kind()
-                    .create_table_caching_triggers_for_table_sql(&source_table)?
-                    .join(";\n");
-                self.execute_batch(&sql).await?;
-                // Add a trigger to clean entries from the cache for the view:
-                let sql = self
-                    .kind()
-                    .create_table_caching_triggers_for_view_sql(&source_table, &view)?
-                    .join(";\n");
-                self.execute_batch(&sql).await?;
-                // Add an entry for the source table triggers to the metacache. If there is another
-                // entry for this source table it will be overwritten, which is desirable in
-                // case it was not previously known if the table was the source table for a view.
-                let mut cache = get_meta_cache()?;
-                let source_triggers_name = format!("{source_table}_triggers");
-                cache.insert(source_triggers_name);
-            }
-            let mut cache = get_meta_cache()?;
-            cache.insert(view_triggers_name);
-        }
-        Ok(())
-    }
-
-    /// Parse the given semi-colon-separated SQL commands and determine which tables will be
-    /// affected (either edited or dropped) by the commands, then ensure that there are no
-    /// entries for those tables in the cache in accordance with the current [CachingStrategy].
-    async fn clear_cache_for_affected_tables(&self, sql: &str) -> Result<(), DbError> {
-        if self.get_caching_strategy() != CachingStrategy::None {
-            let (edited_tables, dropped_tables): (Vec<_>, Vec<_>) = {
-                let (edited_tables, dropped_tables) = get_affected_tables(sql)?;
-                (
-                    edited_tables.into_iter().collect(),
-                    dropped_tables.into_iter().collect(),
-                )
-            };
-            if !edited_tables.is_empty() {
-                let edited_tables: Vec<_> = edited_tables.iter().map(|t| t.as_str()).collect();
-                self.clear_cache_for_edited_tables(&edited_tables).await?;
-            }
-            if !dropped_tables.is_empty() {
-                let dropped_tables: Vec<_> = dropped_tables.iter().map(|t| t.as_str()).collect();
-                self.clear_cache_for_dropped_tables(&dropped_tables).await?;
-            }
-        }
-        Ok(())
-    }
-
-    // Triggers cannot apply to DROP commands, only to INSERT, UPDATE, DELETE, or TRUNCATE.
-    // See https://www.postgresql.org/docs/current/sql-createtrigger.html and
-    // https://sqlite.org/lang_createtrigger.html. Note that PostgreSQL has the concept
-    // of an "event trigger":
-    // https://www.pgtutorial.com/postgresql-tutorial/postgresql-event-triggers/ which could
-    // be used, but SQLite has no such capability. To workaround this limitation, we
-    // define two clear_cache_() functions, one for edited tables, and one for dropped tables.
-    // In the case of a dropped table, unlike an edit, we cannot rely on the caching trigger,
-    // when we are using the [CachingStategy::Trigger] strategy, to automatically delete the
-    // entries from the cache for those tables, since those triggers will have beeen dropped
-    // along with the table.
-    // Although strictly speaking, PostgreSQL (which has event triggers) is not subject to this
-    // limitation, for simplicity we will not be creating a PostgreSQL event trigger and we will
-    // use both functions below for both database types.
-
-    /// Update the cache tables, for the given list of tables, using the current [CachingStrategy],
-    /// under the assumption that the tables in the given list have all just been edited (i.e.,
-    /// truncated, deleted from, inserted to, or updated).
-    async fn clear_cache_for_edited_tables(&self, tables: &[&str]) -> Result<(), DbError> {
-        match self.get_caching_strategy() {
-            CachingStrategy::None | CachingStrategy::Trigger => (),
-            CachingStrategy::TruncateAll => {
-                self.update_last_modified_times(tables).await?;
-                self.delete_query_cache_entries(&[]).await?
-            }
-            CachingStrategy::Truncate => {
-                self.update_last_modified_times(tables).await?;
-                self.delete_query_cache_entries(tables).await?
-            }
-            CachingStrategy::Memory(_) => {
-                self.update_last_modified_times(tables).await?;
-                clear_memory_query_cache(tables)?;
-            }
-        };
-        Ok(())
-    }
-
-    /// Update the cache tables for the given list of tables, using the current [CachingStrategy],
-    /// under the assumption that the tables in the given list have all just been dropped.
-    async fn clear_cache_for_dropped_tables(&self, tables: &[&str]) -> Result<(), DbError> {
-        if let CachingStrategy::Memory(_) = self.get_caching_strategy() {
-            self.update_last_modified_times(tables).await?;
-            clear_memory_query_cache(&tables)?;
-        } else {
-            // Do not clear the cache if the dropped tables include the cache tables themselves:
-            if !tables
-                .iter()
-                .any(|table| [QUERY_CACHE_TABLE, TABLE_CACHE_TABLE].contains(table))
-            {
-                match self.get_caching_strategy() {
-                    CachingStrategy::Memory(_) => unreachable!(),
-                    CachingStrategy::None => (),
-                    CachingStrategy::TruncateAll => {
-                        self.update_last_modified_times(tables).await?;
-                        self.delete_query_cache_entries(&[]).await?;
-                    }
-                    CachingStrategy::Trigger | CachingStrategy::Truncate => {
-                        self.update_last_modified_times(tables).await?;
-                        self.delete_query_cache_entries(tables).await?;
-                    }
-                }
-            }
-        }
-        // Update the meta-cache to remove any entries associated with tables that no longer exist:
-        let mut meta_cache = get_meta_cache()?;
-        for table in tables {
-            if *table == QUERY_CACHE_TABLE {
-                meta_cache.remove(QUERY_CACHE_TABLE);
-            } else if *table == TABLE_CACHE_TABLE {
-                meta_cache.remove(TABLE_CACHE_TABLE);
-            } else {
-                meta_cache.remove(&format!("{table}_triggers"));
-                meta_cache.remove(&format!("{table}_VIEW"));
-                meta_cache.remove(&format!("{table}_TABLE"));
-            }
-        }
-        Ok(())
-    }
-
-    /// Update the last verified time of the query cache entry identified by the triple:
-    /// (tables, statement, params).
-    async fn update_last_verified(
-        &self,
-        tables: &[&str],
-        statement: &str,
-        params: &DbParams,
-    ) -> Result<(), DbError> {
-        match self.get_caching_strategy() {
-            CachingStrategy::Memory(_) => {
-                let mut cache = get_memory_query_cache()?;
-                let epoch_now = SystemTime::now()
-                    .duration_since(UNIX_EPOCH)
-                    .map_err(|err| {
-                        DbError::DataError(format!("Error getting epoch time: {err}"))
-                    })?;
-                let mem_key = MemoryQueryCacheKey {
-                    tables: format!(
-                        "[{}]",
-                        tables
-                            .iter()
-                            .map(|table| format!("\"{table}\""))
-                            .collect::<Vec<_>>()
-                            .join(", ")
-                    ),
-                    statement: statement.to_string(),
-                    parameters: format!("{params:?}"),
-                };
-                match cache.get_mut(&mem_key) {
-                    Some(value) => value.last_verified = epoch_now.as_millis(),
-                    None => (),
-                };
-            }
-            _ => match self.table_exists(QUERY_CACHE_TABLE).await? {
-                true => {
-                    let tables_param = format!(
-                        "[{}]",
-                        tables
-                            .iter()
-                            .map(|table| format!("\"{table}\""))
-                            .collect::<Vec<_>>()
-                            .join(", ")
-                    );
-                    self.execute(
-                        &format!(
-                            r#"UPDATE "{QUERY_CACHE_TABLE}"
-                               SET "last_verified" = {ts}
-                               WHERE "tables" = {p}1
-                               AND "statement" = {p}2
-                               AND "parameters" = {p}3"#,
-                            p = self.kind().param_prefix(),
-                            ts = self.kind().get_epoch_time_sql(),
-                        ),
-                        &[
-                            &format!("[{tables_param}]"),
-                            statement,
-                            &format!("{params:?}"),
-                        ],
-                    )
-                    .await?;
-                }
-                false => (),
-            },
-        };
-        Ok(())
-    }
-
-    /// Update the last modified times of each of the given tables in the table cache.
-    async fn update_last_modified_times(&self, tables: &[&str]) -> Result<(), DbError> {
-        match self.get_caching_strategy() {
-            CachingStrategy::Memory(_) => {
-                let mut cache = get_memory_table_cache()?;
-                let epoch_now = SystemTime::now()
-                    .duration_since(UNIX_EPOCH)
-                    .map_err(|err| {
-                        DbError::DataError(format!("Error getting epoch time: {err}"))
-                    })?;
-                for table in tables {
-                    cache.insert(table.to_string(), epoch_now.as_millis());
-                }
-            }
-            _ => {
-                if self.table_exists(TABLE_CACHE_TABLE).await? {
-                    for table in tables {
-                        let sql = format!(
-                            r#"INSERT INTO "{TABLE_CACHE_TABLE}" ("table", "last_modified")
-                               VALUES ({prefix}1, {ts})
-                               ON CONFLICT ("table") DO UPDATE SET "last_modified" = {ts}"#,
-                            prefix = self.kind().param_prefix(),
-                            ts = self.kind().get_epoch_time_sql(),
-                        );
-                        self.execute_no_cache_clean(&sql, params![table]).await?;
-                    }
-                }
-            }
-        };
-        Ok(())
-    }
-
-    /// Delete the entries for the tables in the given list (independently of the current
-    /// caching strategy) from the query cache table, if it exists. If the given list is empty,
-    /// clear the entire query cache table.
-    async fn delete_query_cache_entries(&self, tables: &[&str]) -> Result<(), DbError> {
-        if self.table_exists(QUERY_CACHE_TABLE).await? {
-            if tables.is_empty() {
-                self.execute_no_cache_clean(&format!(r#"DELETE FROM "{QUERY_CACHE_TABLE}""#), ())
-                    .await?;
-            } else {
-                for table in tables {
-                    let table_param = format!(r#"%"{table}"%"#);
-                    self.execute_no_cache_clean(
-                        &format!(
-                            r#"DELETE FROM "{QUERY_CACHE_TABLE}" WHERE "tables" LIKE {}1"#,
-                            self.kind().param_prefix()
-                        ),
-                        &[table_param],
-                    )
-                    .await?;
-                }
-            }
-        }
-        Ok(())
-    }
-
-    /// Uses the current caching strategy to clear the query cache for any of the given tables
-    /// that (a) are views and (b) have source tables that have been modified more recently than
-    /// the view. This function works both with database and memory cache strategies.
-    async fn update_cached_views(&self, tables: &[&str]) -> Result<(), DbError> {
-        let views = self.which_are_views(tables).await?;
-        match self.get_caching_strategy() {
-            CachingStrategy::Memory(_) => {
-                for view in &views {
-                    let last_verified = {
-                        let mut last_verified = 0;
-                        for (key, value) in get_memory_query_cache()?.iter() {
-                            if key.tables.contains(view) && value.last_verified > last_verified {
-                                last_verified = value.last_verified;
-                            }
-                        }
-                        last_verified
-                    };
-                    let view_sql = self.get_view_sql(&view).await?;
-                    let view_tables = get_view_tables(&view_sql)?;
-                    let last_modified = {
-                        let mut latest_last_modified = 0;
-                        for view_table in &view_tables {
-                            match get_memory_table_cache()?.get(view_table) {
-                                Some(lm) if *lm > latest_last_modified => {
-                                    latest_last_modified = *lm;
-                                }
-                                _ => (),
-                            };
-                        }
-                        latest_last_modified
-                    };
-                    if last_modified >= last_verified {
-                        clear_memory_query_cache(&[view])?;
-                    }
-                }
-            }
-            _ => {
-                for view in &views {
-                    let last_verified = self.last_verified(&view).await?;
-                    let view_sql = self.get_view_sql(&view).await?;
-                    let view_tables = get_view_tables(&view_sql)?;
-                    let last_modified = self
-                        .get_latest_last_modified(
-                            &view_tables.iter().map(|t| t.as_str()).collect::<Vec<_>>(),
-                        )
-                        .await?;
-                    if last_modified >= last_verified {
-                        self.delete_query_cache_entries(&[&view]).await?;
-                    }
-                }
-            }
-        };
-        Ok(())
-    }
-
-    /// Returns the latest of the last modified times of the given tables in the table cache.
-    async fn get_latest_last_modified(&self, tables: &[&str]) -> Result<u64, DbError> {
-        match self.table_exists(TABLE_CACHE_TABLE).await? {
-            true => {
-                let prefix = self.kind().param_prefix().to_string();
-                let mut placeholders = vec![];
-                let mut parameters = vec![];
-                for (i, table) in tables.iter().enumerate() {
-                    let i = i + 1;
-                    placeholders.push(format!("{prefix}{i}"));
-                    parameters.push(DbValue::from(*table));
-                }
-                let placeholders = placeholders.join(",");
-
-                let sql = format!(
-                    r#"SELECT "last_modified"
-                       FROM "{TABLE_CACHE_TABLE}"
-                       WHERE "table" IN ({placeholders})
-                       ORDER BY "last_modified" DESC
-                       LIMIT 1"#,
-                );
-                let rows = self.query_no_cache_clean(&sql, parameters).await?;
-                match rows.len() {
-                    0 => Ok(0),
-                    1 => {
-                        let row = &rows[0];
-                        match row.get("last_modified") {
-                            Some(last_modified) => Ok(last_modified.try_into()?),
-                            None => Err(DbError::DataError(format!(
-                                "No field 'last_modified' in row {row:?}"
-                            ))),
-                        }
-                    }
-                    too_many => Err(DbError::DataError(format!(
-                        "Too many rows returned: {too_many} from table {TABLE_CACHE_TABLE}"
-                    ))),
-                }
-            }
-            false => Ok(0),
-        }
-    }
-
-    /// Gets the last time the given table was modified, as read from the table cache table.
-    /// If there is no entry for the table in the table cache, or if the table cache does not
-    /// exist, returns 0.
-    async fn last_modified(&self, table: &str) -> Result<u64, DbError> {
-        self.get_latest_last_modified(&[table]).await
-    }
-
-    /// Gets the last time that the given table was verified, as read from the query cache table.
-    /// If there is no entry involving the given table in the query cache, or if the query cache
-    /// table doesn't exist, returns 0.
-    async fn last_verified(&self, table: &str) -> Result<u64, DbError> {
-        match self.table_exists(QUERY_CACHE_TABLE).await? {
-            true => {
-                let sql = format!(
-                    r#"SELECT MAX("last_verified") AS "last_verified"
-                       FROM "{QUERY_CACHE_TABLE}"
-                       WHERE "tables" LIKE {p}1"#,
-                    p = self.kind().param_prefix(),
-                );
-                let table_param = format!(r#"%"{table}"%"#);
-                let rows = self.query_no_cache_clean(&sql, &[&table_param]).await?;
-                match rows.first() {
-                    Some(row) => match row.get("last_verified") {
-                        Some(value) if value == DbValue::Null => Ok(0),
-                        Some(value) => Ok(value.try_into()?),
-                        None => Err(DbError::DataError(format!(
-                            "No 'last_verified' found in row: {row:?}"
-                        ))),
-                    },
-                    None => Ok(0),
-                }
-            }
-            false => Ok(0),
-        }
-    }
 
     /// Execute the given SQL command with the given parameters, returning a vector of rows.
     /// If the result of the command exists in the query cache for the given tables, get the
@@ -638,7 +188,7 @@ pub trait DbQuery {
                     // Only views need to be verified every time they are accessed. Tables
                     // do not because they do not have any dependencies.
                     if self.which_are_views(tables).await?.len() > 0 {
-                        self.update_last_verified(tables, sql, &params).await?;
+                        update_last_verified(&self.pool(), tables, sql, &params).await?;
                     }
                     Ok(DbRows { content: db_rows })
                 }
@@ -698,7 +248,7 @@ pub trait DbQuery {
                     // Only views need to be verified every time they are accessed. Tables
                     // do not because they do not have any dependencies.
                     if self.which_are_views(tables).await?.len() > 0 {
-                        self.update_last_verified(tables, sql, &params).await?;
+                        update_last_verified(&self.pool(), tables, sql, &params).await?;
                     }
                     Ok(DbRows { content: db_rows })
                 }
@@ -737,8 +287,8 @@ pub trait DbQuery {
                 Ok(rows)
             }
             CachingStrategy::TruncateAll | CachingStrategy::Truncate => {
-                self.ensure_cache_tables_exist().await?;
-                self.update_cached_views(tables).await?;
+                ensure_cache_tables_exist(&self.pool()).await?;
+                update_cached_views(&self.pool(), tables).await?;
                 let rows = db_cache(tables, sql, &params.into_db_params()).await?;
                 Ok(rows)
             }
@@ -755,26 +305,27 @@ pub trait DbQuery {
                     .difference(&views.iter().map(|v| v.as_str()).collect::<HashSet<_>>())
                     .collect::<Vec<_>>()
                 {
-                    self.ensure_caching_triggers_exist_for_table(&table).await?;
+                    ensure_caching_triggers_exist_for_table(&self.pool(), &table).await?;
                 }
                 for view in &views {
-                    self.ensure_caching_triggers_exist_for_view(view).await?;
+                    ensure_caching_triggers_exist_for_view(&self.pool(), view).await?;
                 }
                 let rows = db_cache(&tables, sql, &params.into_db_params()).await?;
                 Ok(rows)
             }
             CachingStrategy::Memory(cache_size) => {
-                self.update_cached_views(tables).await?;
+                update_cached_views(&self.pool(), tables).await?;
                 let rows = mem_cache(tables, sql, &params.into_db_params(), cache_size).await?;
                 Ok(rows)
             }
         }
     }
 
-    /////////////////////////////////
-
     /// Get the kind of SQL database. See [DbKind] for the supported database kinds.
     fn kind(&self) -> DbKind;
+
+    /// Get the underlying database pool
+    fn pool(&self) -> AnyPool;
 
     /// Get the SQL code that is used to define the given view.
     async fn get_view_sql(&self, view: &str) -> Result<String, DbError> {
@@ -858,7 +409,7 @@ pub trait DbQuery {
 
     /// Execute the given SQL command with the given parameters, returning nothing, and without
     /// updating the cache, regardless of whether the cache-aware-query option
-    /// (see [crate::cache::DbQuery::set_cache_aware_query()]) has been set.
+    /// (see [DbQuery::set_cache_aware_query()]) has been set.
     async fn execute_no_cache_clean(
         &self,
         sql: &str,
@@ -884,7 +435,7 @@ pub trait DbQuery {
 
     /// Execute the given SQL command using the given parameters, returning a vector of rows,
     /// without updating the cache, regardless of whether the cache-aware-query option
-    /// (see [crate::cache::DbQuery::set_cache_aware_query()]) has been set.
+    /// (see [DbQuery::set_cache_aware_query()]) has been set.
     fn query_no_cache_clean(
         &self,
         sql: &str,
