+++ src/rusqlite.rs
@@ -1,7 +1,8 @@
 //! [rusqlite](<https://crates.io/crates/deadpool-sqlite>) implementation for rltbl_db.
 
 use crate::{
-    cache::CachingStrategy,
+    any::AnyPool,
+    cache::{CachingStrategy, clear_cache_for_affected_tables, clear_cache_for_dropped_tables},
     core::{DbError, DbQuery},
     db_kind::{DbKind, MAX_PARAMS_SQLITE},
     db_value::{DbParams, DbRow, DbRows, DbValue, IntoDbParams, IntoDbRows, JsonValue},
@@ -220,6 +221,15 @@ impl DbQuery for RusqlitePool {
         DbKind::SQLite
     }
 
+    /// TODO: Add docstring.
+    fn pool(&self) -> AnyPool {
+        AnyPool::Rusqlite(RusqlitePool {
+            pool: self.pool.clone(),
+            caching_strategy: self.caching_strategy,
+            cache_aware_query: self.cache_aware_query,
+        })
+    }
+
     /// Implements [DbQuery::execute_batch()] for SQLite.
     async fn execute_batch(&self, sql: &str) -> Result<(), DbError> {
         let conn = self
@@ -241,7 +251,7 @@ impl DbQuery for RusqlitePool {
             Ok(_) => {
                 // We need to drop conn here to ensure that any changes to the db are persisted.
                 drop(conn);
-                self.clear_cache_for_affected_tables(sql).await?;
+                clear_cache_for_affected_tables(&self.pool(), sql).await?;
                 Ok(())
             }
         }
@@ -251,7 +261,7 @@ impl DbQuery for RusqlitePool {
     async fn query(&self, sql: &str, params: impl IntoDbParams + Send) -> Result<DbRows, DbError> {
         let rows = self.query_no_cache_clean(sql, params).await?;
         if self.get_cache_aware_query() {
-            self.clear_cache_for_affected_tables(sql).await?;
+            clear_cache_for_affected_tables(&self.pool(), sql).await?;
         }
         Ok(rows)
     }
@@ -426,7 +436,7 @@ impl DbQuery for RusqlitePool {
             .await?;
 
         // Delete dirty entries from the cache in accordance with our caching strategy:
-        self.clear_cache_for_dropped_tables(&[&table]).await?;
+        clear_cache_for_dropped_tables(&self.pool(), &[&table]).await?;
         Ok(())
     }
 
@@ -438,7 +448,7 @@ impl DbQuery for RusqlitePool {
             .await?;
 
         // Delete dirty entries from the cache in accordance with our caching strategy:
-        self.clear_cache_for_dropped_tables(&[&view]).await?;
+        clear_cache_for_dropped_tables(&self.pool(), &[&view]).await?;
         Ok(())
     }
