+++ src/tokio_postgres.rs
@@ -1,7 +1,8 @@
 //! [tokio-postgres](<https://crates.io/crates/deadpool-postgres>) implementation for rltbl_db.
 
 use crate::{
-    cache::CachingStrategy,
+    any::AnyPool,
+    cache::{CachingStrategy, clear_cache_for_affected_tables, clear_cache_for_dropped_tables},
     core::{DbError, DbQuery},
     db_kind::{DbKind, MAX_PARAMS_POSTGRES},
     db_value::{DbParams, DbRow, DbRows, DbValue, IntoDbParams, IntoDbRows, JsonValue},
@@ -219,6 +220,15 @@ impl DbQuery for TokioPostgresPool {
         DbKind::PostgreSQL
     }
 
+    /// TODO: Add docstring.
+    fn pool(&self) -> AnyPool {
+        AnyPool::TokioPostgres(TokioPostgresPool {
+            pool: self.pool.clone(),
+            caching_strategy: self.caching_strategy,
+            cache_aware_query: self.cache_aware_query,
+        })
+    }
+
     /// Implements [DbQuery::execute_batch()] for PostgreSQL
     async fn execute_batch(&self, sql: &str) -> Result<(), DbError> {
         let client =
@@ -230,7 +240,7 @@ impl DbQuery for TokioPostgresPool {
             .await
             .map_err(|err| DbError::DatabaseError(format!("Error in query(): {err:?}")))?;
 
-        self.clear_cache_for_affected_tables(sql).await?;
+        clear_cache_for_affected_tables(&self.pool(), sql).await?;
         Ok(())
     }
 
@@ -238,7 +248,7 @@ impl DbQuery for TokioPostgresPool {
     async fn query(&self, sql: &str, params: impl IntoDbParams + Send) -> Result<DbRows, DbError> {
         let rows = self.query_no_cache_clean(sql, params).await?;
         if self.get_cache_aware_query() {
-            self.clear_cache_for_affected_tables(sql).await?;
+            clear_cache_for_affected_tables(&self.pool(), sql).await?;
         }
         Ok(rows)
     }
@@ -527,7 +537,7 @@ impl DbQuery for TokioPostgresPool {
             .await?;
 
         // Delete dirty entries from the cache in accordance with our caching strategy:
-        self.clear_cache_for_dropped_tables(&[&table]).await?;
+        clear_cache_for_dropped_tables(&self.pool(), &[&table]).await?;
         Ok(())
     }
 
@@ -540,7 +550,7 @@ impl DbQuery for TokioPostgresPool {
             .await?;
 
         // Delete dirty entries from the cache in accordance with our caching strategy:
-        self.clear_cache_for_dropped_tables(&[&view]).await?;
+        clear_cache_for_dropped_tables(&self.pool(), &[&view]).await?;
         Ok(())
     }
