+++ tests/axum.rs
@@ -4,9 +4,7 @@ use axum::{
     response::{Html, IntoResponse},
     routing::get,
 };
-use rltbl_db::{
-    any::AnyPool, cache::DbCache, core::DbQuery, db_kind::DbKind, db_row, db_value::DbRow,
-};
+use rltbl_db::{any::AnyPool, core::DbQuery, db_kind::DbKind, db_row, db_value::DbRow};
 use std::{marker::Sync, sync::Arc};
 use tower_service::Service;
